feat: Update Navbar for improved responsiveness and update Home page layout with new sections and components
